[serve] Move route determination test into test_get_asgi_app_route
